clone: support signature verification
The merge operation (and as a consequence, pull) has had support for
signature verification for quite some time.  However, there were no
support for verifying GPG signatures for the initial clone.

Without signature verification for clone operations, users are forced to
checkout a repository before using verify-commit or verify-tag on the
tree.  This is potentially problematic for a few reasons:

- It's possible to forget to verify the tree when there's nothing that
  enforce signature verification before a cloned repository is used.

- Software on the users system might process the tree before it has been
  successfully verified -- for example, file managers that automagically
  creates thumbnails for images, videos and documents.

Now, this could be worked around with a --no-checkout clone followed by
signature verification and checkout.  There are also various scripts
floating around that more-or-less re-implements 'git clone' with
signature verification by using a combination of init + remote add +
fetch + verify-commit + merge [1].  But none of these options are
particularly user-friendly.

This patch implements signature verification for the clone builtin to
accommodate use-cases where the tree should be verified before use.

There is one major quirk in this patch; namely, recursive clones of
submodules.  It is worked around by passing --no-verify-signatures in
clone_submodule().  The rationale for this approach is that:

- The object ID for submodules are stored in the super repository.
  Thus, if the super is trusted and signed (and the hash function is
  secure), then it seems reasonable to also trust the submodules
  referenced in the super repository.

- Propagating signature verification to recursive clones of submodules
  would lessen the use of signature verification for clones, because
  then users would need the keys for each submodule author in their
  keyring.  Also, not all submodule refs may be signed.

[1] https://gist.github.com/tribut/50c0f7d0b8341fa6d1784c317d5275f0

Signed-off-by: Hans Jerry Illikainen <hji@dyntopia.com>